Make it easier to run program with real input
Previously had to read file and type a long command to get the answer
for real input:

`AdventOfCode.Helpers.read_input("01") |> AdventOfCode.Day01.part1()`

Now there is a macro that injects a new functions `part1/0` and
`part2/0`. When called they will read the file and pass it to `part1/1`
and `part2/2`.

Also removed helper module as it is no longer needed.